Added websocket message that send all characaters position alive

+++ server/src/main/java/my/plaground/Domain/WebSocketParams.java
@@ -9,6 +9,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class WebSocketParams {
     private int characterId;
-    private float positionX;
-    private float positionY;
+    private double positionX;
+    private double positionY;
 }
